pyproject.toml
[project]
name = "library-management"
version = "0.1.0"
description = "A fast api project to manage library application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql==0.3.2",
    "alembic==1.18.1",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "cryptography==46.0.3",
    "databases==0.9.0",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi-cli==0.0.20",
    "fastapi-cloud-cli==0.11.0",
    "fastapi[standard]==0.128.0",
    "fastar==0.8.0",
    "greenlet==3.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mysql-connector-python==9.5.0",
    "pip==24.3.1",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-extra-types==2.11.0",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pymysql==1.1.2",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.21",
    "pyyaml==6.0.3",
    "rich==14.2.0",
    "rich-toolkit==0.17.1",
    "rignore==0.7.6",
    "sentry-sdk==2.49.0",
    "shellingham==1.5.4",
    "sqlalchemy==2.0.45",
    "starlette==0.50.0",
    "typer==0.21.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uvicorn==0.40.0",
    "watchfiles==1.1.1",
    "websockets==16.0",
    "bcrypt",
    "authlib>=1.6.6",
    "python-jose",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
]
